requirements.txt
boto3>=1.26.0
opencv-python-headless>=4.7.0
pillow>=9.4.0
numpy>=1.24.2
requests>=2.28.2
python-dotenv>=1.0.0
